feat: Modernize C code for build compatibility
This commit addresses several issues to modernize the C codebase, ensuring it compiles cleanly on modern systems like Ubuntu 24.04 with GCC.

Key changes include:
- Fixed K&R style function definition in ischeck.c.
- Corrected misuse of logical operators (&&) for bitwise flag checking in vbnodememio.c.
- Resolved operator precedence issues in vblowlevel.c.
- Addressed various compiler warnings:
  - -Wconstant-conversion and -Wshift-count-overflow in vbdecimal.h, vbkeysio.c by using appropriate integer types and literals (INT_MAX, 1ULL).
  - -Wformat in vbmemio.c by using the correct format specifier (%zu) for size_t.
- Removed all instances of the deprecated `register` keyword from byteswap.h.
- Cleaned up code by removing unused variables in isrecover.c.
- Updated build system files by running autoreconf.

+++ libvbisam/vbkeysio.c
@@ -682,7 +682,7 @@ vvbkeyvalueset (const int ihigh, struct keydesc *pskeydesc, unsigned char *pckey
 		case LONGTYPE:
 			iremainder = pskptr->kp_leng;
 			while (iremainder > 0) {
-				inl_stlong (ihigh ? LONG_MAX : LONG_MIN, pckeyvalue);
+				inl_stlong (ihigh ? INT_MAX : INT_MIN, pckeyvalue);
 				pckeyvalue += LONGSIZE;
 				iremainder -= LONGSIZE;
 			}
